feat: Put in place more advanced structured directories
