--- LivingParisApp.Core/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace LivingParisApp.Core;
-
-public class Class1
-{
-
-}
