+++ LivingParisApp.Core/Engines/GraphColoration/ChuLiuEdmonds.cs
@@ -1,0 +1,5 @@
+namespace LivingParisApp.Core.Engines.GraphColoration {
+    public class ChuLiuEdmonds{
+        
+    }
+}
