+++ LivingParisApp.Core/Engines/GraphColoration/WelshPowell.cs
@@ -1,0 +1,5 @@
+namespace LivingParisApp.Core.Engines.GraphColoration {
+    public class WelshPowell{
+        
+    }
+}
