+++ LivingParisApp.Core/Engines/PathFinding/BellmanFord.cs
@@ -1,0 +1,5 @@
+namespace LivingParisApp.Core.Engines.ShortestPaths{
+    public class BellmanFord{
+        
+    }
+}
