+++ LivingParisApp.Core/GraphStructure/Graph.cs
@@ -1,0 +1,5 @@
+namespace LivingParisApp.Core.GraphStructure{
+    public class Graph{
+        
+    }
+}
