+++ LivingParisApp.Core/GraphStructure/Link.cs
@@ -1,0 +1,5 @@
+namespace LivingParisApp.Core.GraphStructure{
+    public class Link{
+        
+    }
+}
