+++ LivingParisApp.Core/GraphStructure/Node.cs
@@ -1,0 +1,5 @@
+namespace LivingParisApp.Core.GraphStructure{
+    public class Node{
+        
+    }
+}
