+++ LivingParisApp.Services/FileHandling/FileReader.cs
@@ -1,0 +1,5 @@
+namespace LivingParisApp.Services.FileHandling {
+    public class FileReader{
+        
+    }
+}
